5794 tempo detection speed improvement

+++ libraries/lib-music-information-retrieval/tests/WavMirAudioReader.cpp
@@ -32,7 +32,7 @@ WavMirAudioReader::WavMirAudioReader(
          mutableSamples.begin());
 }
 
-int WavMirAudioReader::GetSampleRate() const
+double WavMirAudioReader::GetSampleRate() const
 {
    return mSampleRate;
 }
